moves long-running tasks api-models and base to models_library
